Fix - Refactor names so that cases match
