+++ client/src/assets/logos/tsc/Orange_Pin.png
